+++ src/components/Icons/iconChangelogData.ts
@@ -15,6 +15,7 @@ export const changelogData: ChangelogData[] = [
     type: "Update",
     affects: ["Accessibility", "Functionality"],
     notes: [
+      "Added the `navigationArrowDown`, `contentBookmarkOutline`, and `contentBookmark` icons.",
       "Updated the following icon options to remove the ID attributes: `actionHome`, `actionInfo`, `actionList`, `actionPayment`, `actionSearch`, `contentCopy`, `navigationApps`, `navigationMoreVert`, `navigationSubdirectoryArrowLeft`, and `navigationSubdirectoryArrowRight`.",
     ],
   },